Ensure IObservable extensions classes work in the global namespace
Extension classes were defined in their own namespace section containing
custom `using` directives to find IObservable extension methods.

However if a variant was declared in the global namespace then this
class would also be generated in the global namesapce and the orphaned
`using` directives would then break the build as they were no longer at
the top of the file.

To solve this replace all uses of extension method syntax with fully
qualified calls to the System.Reactive extension classes and ditch the
`using`s.