Refactor internal access to unions to prepare for generics
For generic variants there can be multiple types with the same name but
different number of generic arguments, for example

Foo<T1>
Foo<T1, T2>

As such we can no longer rely on the uniqueness of the name when
generating new types. To circumvent this have pre-defined types in the
runtime support library so they can be used as the access key. That now
means we can keep the union types private to the variant implementation,
which has the same generic parameters as the primary type and is thus
unique again.